RemoveSetMethodsMethodCallRector: move tests and classes to new location
