+++ rules-tests/PHPUnit100/Rector/MethodCall/RemoveSetMethodsMethodCallRector/RemoveSetMethodsMethodCallRectorTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\PHPUnit\Tests\CodeQuality\Rector\MethodCall\RemoveSetMethodsMethodCallRector;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\MethodCall\RemoveSetMethodsMethodCallRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
