+++ rules-tests/PHPUnit100/Rector/MethodCall/RemoveSetMethodsMethodCallRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\MethodCall\RemoveSetMethodsMethodCallRector;
+use Rector\PHPUnit\PHPUnit100\Rector\MethodCall\RemoveSetMethodsMethodCallRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(RemoveSetMethodsMethodCallRector::class);
